requirements.txt
python-dotenv
tinkoff-investments
PyQt6
pyqtgraph
matplotlib
numpy